+++ framework/Furion.Pure/ConfigurableOptions/Attributes/MapSettingsAttribute.cs
@@ -1,0 +1,30 @@
+﻿// Copyright (c) 2020-2022 百小僧, Baiqian Co.,Ltd.
+// Furion is licensed under Mulan PSL v2.
+// You can use this software according to the terms and conditions of the Mulan PSL v2.
+// You may obtain a copy of Mulan PSL v2 at:
+//             https://gitee.com/dotnetchina/Furion/blob/master/LICENSE
+// THIS SOFTWARE IS PROVIDED ON AN "AS IS" BASIS, WITHOUT WARRANTIES OF ANY KIND, EITHER EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO NON-INFRINGEMENT, MERCHANTABILITY OR FIT FOR A PARTICULAR PURPOSE.
+// See the Mulan PSL v2 for more details.
+
+namespace Furion.ConfigurableOptions;
+
+/// <summary>
+/// 重新映射属性配置
+/// </summary>
+[SuppressSniffer, AttributeUsage(AttributeTargets.Property | AttributeTargets.Field, AllowMultiple = false)]
+public class MapSettingsAttribute : Attribute
+{
+    /// <summary>
+    /// 构造函数
+    /// </summary>
+    /// <param name="path">appsetting.json 对应键</param>
+    public MapSettingsAttribute(string path)
+    {
+        Path = path;
+    }
+
+    /// <summary>
+    /// 对应配置文件中的路径
+    /// </summary>
+    public string Path { get; set; }
+}