+++ samples/Furion.Application/Startup.cs
@@ -1,6 +1,12 @@
-﻿namespace Furion.Application;
+﻿using Microsoft.Extensions.DependencyInjection;
+
+namespace Furion.Application;
 
 [AppStartup(900)]
 public sealed class Startup : AppStartup
 {
+    public void ConfigureServices(IServiceCollection services)
+    {
+        services.AddConfigurableOptions<AppInfoOptions>();
+    }
 }